+++ src/bot.js
@@ -1,0 +1,37 @@
+const fs = require('fs');
+const Discord = require('discord.js');
+const { prefix, token } = require('./config.json');
+
+const client = new Discord.Client();
+client.commands = new Discord.Collection();
+
+const commandFiles = fs.readdirSync('./commands').filter(file => file.endsWith('.js'));
+
+for (const file of commandFiles) {
+    const command = require(`./commands/${file}`);
+    client.commands.set(command.name, command);
+}
+
+client.on('ready', () => {
+    console.log("Connected as " + client.user.tag);
+})
+
+client.on('message', (message) => {
+    // Preventing bot from replying to another bot or not a command
+    if (!message.content.startsWith(prefix) || message.author.bot) return;
+
+    const args = message.content.slice(prefix.length).split(/ +/);
+    const command = args.shift().toLowerCase();
+
+    // A command was issued but not a valid one
+    if (!client.commands.has(command)) return;
+
+    try {
+        client.commands.get(command).execute(message, args);
+    } catch (error) {
+        console.log(error);
+        message.reply('It looks like something went horribly wrong! Sorry.');
+    }
+})
+
+client.login(token);